replace registry with v2 api
